requirements.txt
anyio==4.6.2.post1
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
beautifulsoup4==4.12.3
blinker==1.9.0
Brotli==1.1.0
certifi==2024.7.4
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==44.0.1
cssselect2==0.7.0
diff-match-patch==20241021
distlib==0.3.8
dj-database-url==2.2.0
Django==5.1.3
django-anymail==13.1
django-autoslug==1.9.9
django-cities-light==3.10.2
django-countries==7.6.1
django-debug-toolbar==4.4.6
django-environ==0.11.2
django-extensions==4.1
django-import-export==4.3.9
django-phonenumber-field==8.0.0
django-smart-selects==1.7.2
django-unfold==0.65.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.1
feedparser==6.0.11
filelock==3.15.4
fonttools==4.56.0
geographiclib==2.0
geopy==2.4.1
git-filter-repo==2.45.0
gunicorn==22.0.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httpx==0.27.2
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
lxml==6.0.1
MarkupSafe==3.0.2
oscrypto==1.3.0
packaging==24.0
phonenumbers==8.13.54
pillow==10.4.0
platformdirs==4.2.2
progressbar2==4.5.0
psycopg==3.2.10
psycopg-binary==3.2.10
pycparser==2.22
pydyf==0.11.0
pyHanko==0.25.3
pyhanko-certvalidator==0.26.5
pypdf==5.3.0
pyphen==0.17.2
python-barcode==0.16.1
python-bidi==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-utils==3.9.1
pytube==15.0.0
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
reportlab==4.3.0
requests==2.32.3
sgmllib3k==1.0.0
simplejson==3.19.3
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.1
stripe==11.1.1
svglib==1.5.1
sweetify==2.3.1
tablib==3.8.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.0
tzdata==2024.1
tzlocal==5.2
Unidecode==1.4.0
uritools==4.0.3
urllib3==2.2.2
virtualenv==20.26.3
weasyprint==64.0
webencodings==0.5.1
Werkzeug==3.1.3
xhtml2pdf==0.2.16
zopfli==0.2.3.post1
